refactor: Mene 패키지에서 Product 분리
