+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -7,11 +7,11 @@ import java.util.stream.Collectors;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.menu.exception.NotFoundException;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.menu.repository.ProductRepository;
+import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.menu.ui.dto.request.MenuCreateRequest;
 import kitchenpos.menu.ui.dto.request.MenuProductCreateRequest;
 import kitchenpos.menu.ui.dto.response.MenuCreateResponse;
