+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -7,7 +7,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.menu.domain.common.Price;
 
 @Entity
 public class Menu {
