+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -7,6 +7,7 @@ import javax.persistence.ElementCollection;
 import javax.persistence.Embeddable;
 import javax.persistence.JoinColumn;
 import kitchenpos.menu.exception.NotFoundException;
+import kitchenpos.product.domain.Product;
 
 @Embeddable
 public class MenuProducts {
