+++ src/main/java/kitchenpos/menu/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain.common;
+package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
 import java.util.Objects;
