+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.repository;
+package kitchenpos.product.repository;
 
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
