+++ src/main/java/kitchenpos/product/ui/dto/response/ProductFindAllResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.menu.ui.dto.response;
+package kitchenpos.product.ui.dto.response;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductFindAllResponse {
