+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -2,7 +2,7 @@ package kitchenpos.menu.application;
 
 import static kitchenpos.menu.fixture.MenuFixture.createMenu;
 import static kitchenpos.menu.fixture.MenuGroupFixture.createMenuGroup;
-import static kitchenpos.menu.fixture.ProductFixture.createProduct;
+import static kitchenpos.product.fixture.ProductFixture.createProduct;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -18,11 +18,11 @@ import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.menu.exception.NotFoundException;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.menu.repository.ProductRepository;
+import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.menu.ui.dto.request.MenuCreateRequest;
 import kitchenpos.menu.ui.dto.request.MenuProductCreateRequest;
 import kitchenpos.menu.ui.dto.response.MenuCreateResponse;
