+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.application;
+package kitchenpos.product.application;
 
-import static kitchenpos.menu.fixture.ProductFixture.createProduct;
+import static kitchenpos.product.fixture.ProductFixture.createProduct;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -8,11 +8,11 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
-import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.repository.ProductRepository;
-import kitchenpos.menu.ui.dto.request.ProductCreateRequest;
-import kitchenpos.menu.ui.dto.response.ProductCreateResponse;
-import kitchenpos.menu.ui.dto.response.ProductFindAllResponse;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.product.ui.dto.request.ProductCreateRequest;
+import kitchenpos.product.ui.dto.response.ProductCreateResponse;
+import kitchenpos.product.ui.dto.response.ProductFindAllResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
